server configuration added for dev and prod
